Added jcrop app, which allows cropping of images.
The JS code utilizes the jQuery plugin 'Jcrop' to get a window size for the new image after crop effect. The parameters
for the new image are passed to html hidden inputs and processed by a post handler by the server.

The django server reads for the x,y coords and width and height and image url. The image url is fetched and processed
into byte data that is then handled by PIL. After cropping the image, it is returned in a response and saved to an
arbitrary location on the 'MEDIA_ROOT' path.